Move all modules to modules folder
