+++ src/modules/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { JwtModule } from '@nestjs/jwt';
-import { AccountModule } from 'src/account/account.module';
+import { AccountModule } from 'src/modules/account/account.module';
 import { PassportModule } from '@nestjs/passport';
 import { jwtConstants } from './constants';
 import { LocalStrategy } from './local.strategy';
