+++ src/modules/index.ts
@@ -1,0 +1,7 @@
+export * from './account/account.module'
+export * from './auth/auth.module'
+export * from './booking/booking.module'
+export * from './dummy/dummy.module'
+export * from './notification/notification.module'
+export * from './review/review.module'
+export * from './search/search.module'
